feat: implement all work target management and partial my work target
