+++ app/Http/Controllers/WorkTargetController.php
@@ -8,6 +8,7 @@ use App\Models\WorkTargetValue;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Inertia\Inertia;
+use Symfony\Component\Uid\Ulid;
 
 class WorkTargetController extends Controller
 {
@@ -171,58 +172,74 @@ class WorkTargetController extends Controller
 
     public function indexMe(Request $request)
     {
+        $user = $request->user();
+
+        $workTargets = DB::table('work_targets')
+            ->where('work_targets.assigned_id', $user->id)
+            ->select(
+                'work_targets.*',
+            )
+            ->get();
+
         return Inertia::render('my-work-targets/index', [
-            'workTargets' => [],
+            'workTargets' => $workTargets,
         ]);
     }
 
     public function indexKaur(Request $request)
     {
         $role = "kaur";
-
-        $staffs = DB::table('work_target_values')
-            ->rightJoin('users', 'users.id', '=', 'work_target_values.user_id')
-            ->where('users.role', $role)
-            ->select(
-                'users.*',
-                DB::raw('CEIL(COALESCE(AVG(first_quarter_score), 0)) as average_first_quarter_score'),
-                DB::raw('CEIL(COALESCE(AVG(second_quarter_score), 0)) as average_second_quarter_score'),
-                DB::raw('CEIL(COALESCE(AVG(third_quarter_score), 0)) as average_third_quarter_score'),
-                DB::raw('CEIL(COALESCE(AVG(fourth_quarter_score), 0)) as average_fourth_quarter_score'),
-            )
-            ->groupBy('users.id')
-            ->get();
-
         return Inertia::render('work-targets-management/index', [
             'role' => $role,
-            'staffs' => $staffs,
-            'workTargets' => [],
-            'performances' => [],
+            'staffs' => [],
         ]);
     }
 
     public function indexStaf(Request $request)
     {
-        $role = "tpa";
+        $user = $request->user();
+        $role = "staf";
 
-        $staffs = DB::table('work_target_values')
-            ->rightJoin('users', 'users.id', '=', 'work_target_values.user_id')
-            ->where('users.role', $role)
+        // $staffs = DB::table('work_target_values')
+        //     ->rightJoin('users', 'users.id', '=', 'work_target_values.user_id')
+        //     ->where('users.role', $role)
+        //     ->select(
+        //         'users.*',
+        //         DB::raw('CEIL(COALESCE(AVG(first_quarter_score), 0)) as average_first_quarter_score'),
+        //         DB::raw('CEIL(COALESCE(AVG(second_quarter_score), 0)) as average_second_quarter_score'),
+        //         DB::raw('CEIL(COALESCE(AVG(third_quarter_score), 0)) as average_third_quarter_score'),
+        //         DB::raw('CEIL(COALESCE(AVG(fourth_quarter_score), 0)) as average_fourth_quarter_score'),
+        //     )
+        //     ->groupBy('users.id')
+        //     ->get();
+
+        $staffs = DB::table('work_targets')
+            ->rightJoin('users as assigned', 'assigned.id', '=', 'work_targets.assigned_id')
+            ->leftJoin('users as creator', 'creator.id', '=', 'work_targets.creator_id')
+            ->where('assigned.role', $role)
+            ->where('assigned.division', $user->division)
             ->select(
-                'users.*',
-                DB::raw('CEIL(COALESCE(AVG(first_quarter_score), 0)) as average_first_quarter_score'),
-                DB::raw('CEIL(COALESCE(AVG(second_quarter_score), 0)) as average_second_quarter_score'),
-                DB::raw('CEIL(COALESCE(AVG(third_quarter_score), 0)) as average_third_quarter_score'),
-                DB::raw('CEIL(COALESCE(AVG(fourth_quarter_score), 0)) as average_fourth_quarter_score'),
+                'assigned.*',
+                DB::raw('COALESCE(CEIL(AVG(first_quarter_score)), 0) as average_first_quarter_score'),
+                DB::raw('COALESCE(CEIL(AVG(second_quarter_score)), 0) as average_second_quarter_score'),
+                DB::raw('COALESCE(CEIL(AVG(third_quarter_score)), 0) as average_third_quarter_score'),
+                DB::raw('COALESCE(CEIL(AVG(fourth_quarter_score)), 0) as average_fourth_quarter_score'),
             )
-            ->groupBy('users.id')
+            ->groupBy('assigned.id')
             ->get();
+
+        foreach ($staffs as $staff) {
+            $staff->work_targets = DB::table('work_targets')
+                ->where('assigned_id', $staff->id)
+                ->select(
+                    'work_targets.*',
+                )
+                ->get();
+        }
 
         return Inertia::render('work-targets-management/index', [
             'role' => $role,
             'staffs' => $staffs,
-            'workTargets' => [],
-            'performances' => [],
         ]);
     }
 
@@ -243,7 +260,8 @@ class WorkTargetController extends Controller
         try {
             // Validate the request data
             $validatedData = $request->validate([
-                'name' => 'required|string|max:255'
+                'name' => 'required|string|max:255',
+                'assigned_id' => 'required|exists:users,id',
             ]);
 
             $userId = $request->user()->id;
@@ -252,12 +270,14 @@ class WorkTargetController extends Controller
             DB::beginTransaction();
 
             // Create a new work target
-            $workTarget = new WorkTarget();
-            $workTarget->name = $validatedData['name'];
-            $workTarget->user_id = $userId;
-
-            // Save the work target to the database
-            $workTarget->save();
+            DB::table('work_targets')->insert([
+                'id' => Ulid::generate(),
+                'name' => $validatedData['name'],
+                'creator_id' => $userId,
+                'assigned_id' => $validatedData['assigned_id'],
+                'created_at' => now(),
+                'updated_at' => now(),
+            ]);
 
             DB::commit();
 
@@ -312,51 +332,104 @@ class WorkTargetController extends Controller
             // Validate the request data
             $validatedData = $request->validate([
                 'name' => 'required|string|max:255',
-                'staffIds' => 'array',
-                'unit' => 'required|in:week,total,day,minute',
-                'comparator' => 'required|in:eq,lte,gte,gt,lt',
-                'first_quarter_target' => 'required|integer',
-                'second_quarter_target' => 'required|integer',
-                'third_quarter_target' => 'required|integer',
-                'fourth_quarter_target' => 'required|integer',
+                'unit' => 'optional|in:at_week,work_day,total,percentage',
+                'first_quarter_target' => 'optional|integer',
+                'second_quarter_target' => 'optional|integer',
+                'third_quarter_target' => 'optional|integer',
+                'fourth_quarter_target' => 'optional|integer',
             ]);
 
             // Begin a database transaction
             DB::beginTransaction();
 
-            // Find the work target by ID
-            $workTarget = WorkTarget::findOrFail($id);
+            DB::table('work_targets')
+                ->where('id', $id)
+                ->update([
+                    'name' => $validatedData['name'],
+                    'unit' => $validatedData['unit'],
+                    'first_quarter_target' => $validatedData['first_quarter_target'],
+                    'second_quarter_target' => $validatedData['second_quarter_target'],
+                    'third_quarter_target' => $validatedData['third_quarter_target'],
+                    'fourth_quarter_target' => $validatedData['fourth_quarter_target'],
+                    'updated_at' => now(),
+                ]);
 
-            // Update the work target's name
-            $workTarget->name = $validatedData['name'];
-            $workTarget->unit = $validatedData['unit'];
-            $workTarget->comparator = $validatedData['comparator'];
-            $workTarget->first_quarter_target = $validatedData['first_quarter_target'];
-            $workTarget->second_quarter_target = $validatedData['second_quarter_target'];
-            $workTarget->third_quarter_target = $validatedData['third_quarter_target'];
-            $workTarget->fourth_quarter_target = $validatedData['fourth_quarter_target'];
+            // Commit the transaction
+            DB::commit();
 
-            // Save the changes to the database
-            $workTarget->save();
+            return back();
+        } catch (\Exception $e) {
+            // Rollback the transaction if an error occurs
+            DB::rollBack();
 
-            // Delete existing work target values that are not in the request
-            $existingStaffIds = $workTarget->workTargetValues()->pluck('user_id')->toArray();
-            $staffIdsToDelete = array_diff($existingStaffIds, $validatedData['staffIds']);
+            return back();
+        }
+    }
 
-            WorkTargetValue::whereIn('user_id', $staffIdsToDelete)
-                ->where('work_target_id', $workTarget->id)
-                ->delete();
+    public function assess(Request $request, string $id)
+    {
+        try {
+            // Validate the request data
+            $validatedData = $request->validate([
+                'first_quarter_score' => 'required|integer',
+                'second_quarter_score' => 'required|integer',
+                'third_quarter_score' => 'required|integer',
+                'fourth_quarter_score' => 'required|integer',
+                'final_score' => 'required|integer',
+            ]);
 
-            // Create work target values for each staff that is not already in the database
-            foreach ($validatedData['staffIds'] as $staffId) {
-                if (!in_array($staffId, $existingStaffIds)) {
-                    $workTargetValue = new WorkTargetValue();
-                    $workTargetValue->user_id = $staffId;
-                    $workTargetValue->work_target_id = $workTarget->id;
 
-                    $workTargetValue->save();
-                }
-            }
+            // Begin a database transaction
+            DB::beginTransaction();
+
+            DB::table('work_targets')
+                ->where('id', $id)
+                ->update([
+                    'first_quarter_score' => $validatedData['first_quarter_score'],
+                    'second_quarter_score' => $validatedData['second_quarter_score'],
+                    'third_quarter_score' => $validatedData['third_quarter_score'],
+                    'fourth_quarter_score' => $validatedData['fourth_quarter_score'],
+                    'final_score' => $validatedData['final_score'],
+                    'updated_at' => now(),
+                ]);
+
+            // Commit the transaction
+            DB::commit();
+
+            return back();
+        } catch (\Exception $e) {
+            // Rollback the transaction if an error occurs
+            DB::rollBack();
+
+            return back();
+        }
+    }
+
+    public function submit(Request $request, string $id)
+    {
+        try {
+            // Validate the request data
+            $validatedData = $request->validate([
+                'first_quarter_value' => 'required|integer',
+                'second_quarter_value' => 'required|integer',
+                'third_quarter_value' => 'required|integer',
+                'fourth_quarter_value' => 'required|integer',
+                'category' => 'required|in:light,medium,heavy',
+            ]);
+
+            // Begin a database transaction
+            DB::beginTransaction();
+
+            DB::table('work_targets')
+                ->where('id', $id)
+                ->update([
+                    'first_quarter_value' => $validatedData['first_quarter_value'],
+                    'second_quarter_value' => $validatedData['second_quarter_value'],
+                    'third_quarter_value' => $validatedData['third_quarter_value'],
+                    'fourth_quarter_value' => $validatedData['fourth_quarter_value'],
+                    'category' => $validatedData['category'],
+                    'updated_at' => now(),
+                ]);
 
             // Commit the transaction
             DB::commit();
@@ -377,14 +450,12 @@ class WorkTargetController extends Controller
     {
 
         try {
-            // Find the work target by ID
-            $workTarget = WorkTarget::findOrFail($id);
-
             // Begin a database transaction
             DB::beginTransaction();
 
-            // Delete the work target
-            $workTarget->delete();
+            DB::table('work_targets')
+                ->where('id', $id)
+                ->delete();
 
             // Commit the transaction
             DB::commit();
