+++ resources/js/pages/my-work-targets/index.tsx
@@ -31,12 +31,13 @@ import {
 	TableRow,
 } from "../../components/ui/table";
 import DashboardLayout from "../../layouts/dashboard-layout";
-import type { WorkTarget, WorkTargetValue } from "../../types";
+import { WorkTargetCategory, WorkTargetUnit } from "../../lib/enums";
+import type { WorkTarget } from "../../types";
 
 export default function MyWorkTargets({
 	workTargets,
 }: {
-	workTargets: (WorkTarget & WorkTargetValue)[];
+	workTargets: WorkTarget[];
 }) {
 	const [selectedWorkTargetId, setSelectedWorkTargetId] = useState<
 		string | null
@@ -139,7 +140,7 @@ function WorkTargetRow({
 	selectedWorkTargetId,
 	setSelectedWorkTargetId,
 }: {
-	workTarget: WorkTarget & WorkTargetValue;
+	workTarget: WorkTarget;
 	idx: number;
 	selectedWorkTargetId: string | null;
 	setSelectedWorkTargetId: (id: string | null) => void;
@@ -157,7 +158,7 @@ function WorkTargetRow({
 			...values,
 		};
 
-		router.put(`/dashboard/performance/me/detail/${workTarget.id}`, req, {
+		router.post(`/dashboard/work-target/${workTarget.id}/submit`, req, {
 			preserveState: true,
 		});
 		setSelectedWorkTargetId(null);
@@ -168,14 +169,7 @@ function WorkTargetRow({
 			<TableCell className="py-3 px-4">{idx + 1}</TableCell>
 			<TableCell className="py-3 w-full px-4">{workTarget.name}</TableCell>
 			<TableCell className="py-3 px-4 text-center">
-				{
-					{
-						minute: "Menit",
-						day: "Hari",
-						total: "Jumlah",
-						week: "Minggu",
-					}[workTarget.unit]
-				}
+				{WorkTargetUnit[workTarget.unit]}
 			</TableCell>
 			<TableCell className="py-3 px-4 text-center">
 				{workTarget.first_quarter_target}
@@ -284,7 +278,7 @@ function WorkTargetRow({
 						setSelectedWorkTargetId(workTarget.id);
 						setValues((prev) => ({
 							...prev,
-							category: value as "light" | "medium" | "heavy",
+							category: value as keyof typeof WorkTargetCategory,
 						}));
 					}}
 				>
@@ -292,24 +286,16 @@ function WorkTargetRow({
 						<SelectValue placeholder="Ringan" />
 					</SelectTrigger>
 					<SelectContent>
-						{[
-							{
-								label: "Ringan",
-								value: "light",
-							},
-							{
-								label: "Sedang",
-								value: "medium",
-							},
-							{
-								label: "Berat",
-								value: "heavy",
-							},
-						].map((value) => (
-							<SelectItem key={value.value} value={value.value}>
-								{value.label}
-							</SelectItem>
-						))}
+						{Object.entries(WorkTargetCategory)
+							.map(([key, value]) => ({
+								key,
+								value,
+							}))
+							.map((category) => (
+								<SelectItem key={category.key} value={category.key}>
+									{category.value}
+								</SelectItem>
+							))}
 					</SelectContent>
 				</Select>
 			</TableCell>
